Improve logs for all topics
To have more info in logs about the received message,
we logged what we found in message.body["what"].
The problem is that that key is present only in copr topics,
which is the only topics we initially started with.
Since then we added more topics, so we need to check different
keys for each topic if want to log what we actually consumed.